Spherical harmonics

Starting from version 2.4, Healpix.jl implements generalized Fourier transformations through the libsharp library, to convert a map from its pixel-space representation to its decomposition in spherical harmonics. This has multiple applications, the most relevant being the analysis of Cosmic Microwave Background maps and the efficient computation of convolution operators.

Everything revolves around the Alm type, which encodes a set of spherical harmonics and is thus conceptually equivalent to the concept of a HealpixMap, only living in the harmonic space:

Healpix.Alm — Type

Alm{T <: Number, AA <: AbstractVector{T}}

An array of harmonic coefficients (a_ℓm).

The type T is used for the value of each harmonic coefficient, and it must be a Number (one should however only use complex types for this). The type AA is used to store the array of coefficients; a typical choice is Vector.

A Alm type contains the following fields:

alm: the array of harmonic coefficients

lmax: the maximum value for $ℓ$

mmax: the maximum value for $m$

tval: maximum number of $m$ coefficients for the maximum $ℓ$

The $a_{ℓm}$ are stored by $m$: if $ℓ_{max}$ is 16, the first 16 elements are $m=0$, $ℓ=0-16$, then the following 15 elements are $m=1$, $ℓ=1-16$, then $m=2$, $ℓ=2-16$ and so on until the last element, the 153th, is $m=16$, $ℓ=16$.

In the general case, the number of coefficients in a spherical harmonic expansion is infinite. For obvious reasons, Healpix.jl only allows to store band-limited expansions. The function numberOfAlms returns the number of floating-point numbers used to store the expansion, as a function of the maximum value for $\ell$ and $m$.

Healpix.numberOfAlms — Function

numberOfAlms(lmax::Integer, mmax::Integer) -> Integer
numberOfAlms(lmax::Integer) -> Integer

Return the size of the array of complex numbers needed to store the a_ℓm coefficients in the range of ℓ and m specified by lmax and mmax. If mmax is not specified, it is assumed to be equal to lmax. If lmax and mmax are inconsistent or negative, a DomainError exception is thrown.

Converting between pixel space and harmonic space

Healpix.jl implements the four functions alm2map, map2alm, alm2map!, and map2alm! to convert a map from a pixel-space representation to the harmonic space and vice-versa. The functions ending with ! are mutating functions, which means that they assume that the result must be saved in a preallocated variable; they are space- and time-efficient and should be used when you want your code to be performant, or when you plan to apply the same operation several times (e.g., in a Monte Carlo simulation). The synthesis operation (alm2map) is generally referred to with the matrix operator $\mathrm{Y}$, while his inverse (map2alm) with $\mathrm{Y}^{-1}$.

Healpix.jl also implements the two adjoint functions adjoint_alm2map! and adjoint_map2alm!, represented by $\mathrm{Y}^{\mathrm{T}}$ and $(\mathrm{Y}^{-1})^\mathrm{T}$ respectively. While the synthesis operator on a general scalar field $f(θ, φ)$ can be defined through an exact summation as $f(θ, φ) = \mathrm{Y} \, a_{ℓm}$ where $f(θ, φ) = \sum_{ℓ=0}^∞ \sum_{m=-ℓ}^ℓ a_{ℓm} Y_{ℓm}(θ, φ)$. The analysis operator is defined through an integral operator as $a_{ℓm} = \mathrm{Y}^{-1} f(θ, φ)$ where $a_{ℓm} = \int_0^{2π} \int_0^π Y_{ℓm}^*(θ, φ)\, f(θ, φ) \sin θ \, dθ \, dφ$. Though, in the real case wherein maps are pixelized, the latter ends up being approximated through a summation over the pixels. Here is where the adjoint of the synthesis operator, $\mathrm{Y}^{\mathrm{T}}$, comes into play. It is defined through: $\mathrm{Y}^{\mathrm{T}} f(θ, φ) \equiv \sum_{i=1}^{N_{\mathrm{pix}}} Y_{ℓm,\,i}^* \, f_i,$ which is an exact operation. Note that the latter does not give directly the $a_{\ell m}$ coefficients, since $\mathrm{Y}^{-1} \simeq \mathrm{W}\, \mathrm{Y}^{\mathrm{T}}$, where $\mathrm{W}$ is a diagonal matrix whose non-zero elements are approximately constant and equal to $4π / N_{\mathrm{pix}}$, depending on the map pixelization. The latter realtion is also useful to obtain the adjoint of the analysis operator: $(\mathrm{Y}^{-1})^\mathrm{T} = \mathrm{W}^{\mathrm{T}}\,\mathrm{Y} = \mathrm{W}\,\mathrm{Y}$.

Here is an example:

using Random

# Ensure reproducibility by using a fixed seed
Random.seed!(1234)

nside = 8
m = HealpixMap{Float32,RingOrder}(nside)

# Initialize the pixels to random values in the 0…1 range
for i in 1:length(m)
    m[i] = rand(Float32)
end

alm = map2alm(m)

# Go back to pixel space
newm = alm2map(alm, nside)

768-element HealpixMap{Float64, RingOrder, Vector{Float64}}:
 0.4334681110199904
 0.46760588403582776
 0.2356189978218053
 0.7677461691857397
 0.26188278236087537
 0.438393398270254
 0.9302889363247367
 0.8597877691283009
 0.6675362959627755
 0.5961908196786236
 ⋮
 0.40445452823379824
 0.8160867498711573
 0.6332998058370749
 0.612791914793216
 0.4012637841276252
 0.19598229849597731
 0.6998262769237045
 0.8132148132630672
 0.43815195983473854

The variable newm is a map that is close enough to m, yet it is not exactly the same because of the approximations done by both map2alm and alm2map.

Healpix.map2alm! — Function

map2alm!(map::HealpixMap{Float64, RingOrder, Vector{Float64}},
         alm::Alm{ComplexF64, Vector{ComplexF64}};
         niter::Integer=3)
map2alm!(map::PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}},
         alm::Vector{Alm{ComplexF64, Vector{ComplexF64}}};
         niter::Integer=3)

This function performs a spherical harmonic transform on the map and places the results in the passed alm object. This function requires types derived from Float64, since it is done in-place.

Arguments

map: the map that must be decomposed in spherical harmonics. It can either be a HealpixMap{Float64, RingOrder} type (scalar map) or a PolarizedHealpixMap{Float64, RingOrder} type (polarized map).

alm::Alm{ComplexF64, Vector{ComplexF64}}: the spherical harmonic coefficients to be written to.

Keywords

niter::Integer: number of iterations of SHTs to perform, to enhance accuracy

Healpix.map2alm — Function

map2alm(map::HealpixMap{Float64, RingOrder, AA};
        lmax=nothing, mmax=nothing, niter::Integer=3)
map2alm(m::HealpixMap{T, RingOrder, AA};
        lmax=nothing, mmax=nothing,
        niter::Integer=3) where {T <: Real, AA <: AbstractVector{T}}

Compute the spherical harmonic coefficients of a map. To enhance precision, more iterations of the transforms can be performed by passing a nonzero niter. The underlying SHT library libsharp performs all calculations using Cdouble types, so all inputs are converted to types based on Float64.

Arguments

map: the map to decompose in spherical harmonics. It can either be a HealpixMap{T, RingOrder, AA} type (scalar map) or a PolarizedHealpixMap{T, RingOrder, AA} type (polarized map).

Keywords

lmax::Integer: the maximum ℓ coefficient, will default to 3*nside-1 if not specified.

mmax::Integer: the maximum m coefficient

niter::Integer: number of SHT iterations, to enhance precision. Defaults to 3

Returns

Alm{ComplexF64, Vector{ComplexF64}}: the spherical harmonic coefficients corresponding to the map

Healpix.alm2map! — Function

alm2map!(alm::Alm{ComplexF64, Vector{ComplexF64}},
         map::HealpixMap{Float64, RingOrder, Vector{Float64}})
alm2map!(alm::Vector{Alm{ComplexF64, Vector{ComplexF64}}},
         map::PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}})

This function performs a spherical harmonic transform on the map and places the results in the passed alm object. This function requires types derived from Float64, since it is done in-place.

Arguments

alm::Alm{ComplexF64, Vector{ComplexF64}}: the spherical harmonic coefficients to perform the spherical harmonic transform on.

map: the map that will contain the result. It can either be a HealpixMap{Float64, RingOrder, Vector{Float64}} type (scalar map) or a PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}} (polarized map).

Healpix.alm2map — Function

alm2map(alm::Alm{ComplexF64, Vector{Float64}}, nside::Integer)
alm2map(alm::Alm{T, Vector{T}}, nside::Integer) where T
alm2map(alm::Vector{Alm{ComplexF64, Vector{ComplexF64}}}, nside::Integer)
alm2map(alms::Vector{Alm{T, Vector{T}}}, nside::Integer) where T

Compute a map from spherical harmonic coefficients. The underlying SHT library libsharp performs all calculations in Cdouble, so all inputs are converted to types based on Float64.

Arguments

alm: the spherical harmonic coefficients to transform. If of type Alm{T, Vector{T}}, we assume a spin-0 spherical harmonic transform. If an array of Alm is passed, we assume that the components correspond to T, E, and B coefficients.

Keywords

nside::Integer: Healpix resolution parameter

Returns

HealpixMap{Float64, RingOrder, Vector{Float64}} or PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}} depending on if the input alm is of type Alm{T, Vector{T}} or Array{Alm{T,Vector{T}}} respectively.

Healpix.adjoint_alm2map! — Function

adjoint_alm2map!(map::HealpixMap{Float64, RingOrder, Vector{Float64}},
                 alm::Alm{ComplexF64, Vector{ComplexF64}})
adjoint_alm2map!(map::PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}},
                 alm::Vector{Alm{ComplexF64, Vector{ComplexF64}}})

This function performs a spherical harmonic transform Yᵀ on the map and places the results in the passed alm object. This function requires types derived from Float64, since it is done in-place.

Arguments

map: the map that must be decomposed in spherical harmonics. It can either be a HealpixMap{Float64, RingOrder} type (scalar map) or a PolarizedHealpixMap{Float64, RingOrder} type (polarized map).

alm::Alm{ComplexF64, Vector{ComplexF64}}: the spherical harmonic coefficients to be written to.

Healpix.adjoint_map2alm! — Function

adjoint_map2alm!(alm::Alm{ComplexF64, Vector{ComplexF64}},
                 map::HealpixMap{Float64, RingOrder, Vector{Float64}})
adjoint_map2alm!(alm::Vector{Alm{ComplexF64, Vector{ComplexF64}}},
                 map::PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}})

This function performs the spherical harmonic transform (Y^-1)^t = W^t Y = W Y on the map and places the results in the passed alm object. This function requires types derived from Float64, since it is done in-place.

Arguments

alm::Alm{ComplexF64, Vector{ComplexF64}}: the spherical harmonic coefficients to perform the spherical harmonic transform on.

map: the map that will contain the result. It can either be a HealpixMap{Float64, RingOrder, Vector{Float64}} type (scalar map) or a PolarizedHealpixMap{Float64, RingOrder, Vector{Float64}} (polarized map).

Healpix.iterate_map2alm! — Function

iterate_map2alm!(
    maps::Vector{Vector{Float64}}, alms::Vector{Vector{Complex{Float64}}},
    geom_info::Libsharp.GeomInfo, alm_info::Libsharp.AlmInfo,
    niter::Integer, spin::Integer)

This is an internal function for implementing iterative map2alm, used when the parameter niter is greater than 0. It synthesizes a map from the alms, subtracts it from the map to form a residual map, and then adds the harmonic coefficients of the residual map to the alms. It repeats this niter times. It performs this in-place on arrays of Float64 and Complex{Float64}.

Arguments

maps::Vector{Vector{Float64}}: an array where each element is a 1D Healpix map array.

alms::Vector{Vector{Complex{Float64}}}: an array where each element is a array of alms

geom_info::Libsharp.GeomInfo: contains information about the pixelization

alm_info::Libsharp.AlmInfo: contains information about the SHT coefficient ordering

niter::Integer: number of iterations to perform

spin::Integer: spin of the field, 0 or 2

From harmonic coefficients to the power spectrum

You can use the function alm2cl to convert a set of $a_{\ell m}$ coefficients into the components $C_\ell$ of the power spectrum. The pixelization also induces a transfer function, which can be obtained from pixwin. A simple Gaussian beam window function in the asymptotic small-beam limit can be computed with gaussbeam.

Healpix.alm2cl — Function

alm2cl(alm::Alm{Complex{T}}) where {T <: Number} -> Vector{T}
alm2cl(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number} -> Vector{T}

Compute $C_ℓ$ from the spherical harmonic coefficients of one or two fields.

Arguments

alm₁::Alm{Complex{T}}: the spherical harmonic coefficients of the first field

alm₂::Alm{Complex{T}}: the spherical harmonic coefficients of the second field

Returns

Array{T} containing $C_ℓ$, with the first element referring to ℓ=0.

Healpix.pixwin — Function

pixwin(nside; pol=false)

Arguments:

nside: HEALPix resolution parameter

Keywords

pol::Bool=false: if true, also return polarization pixel window

Returns:

Vector{Float64} pixel window function. If pol=true, returns a Tuple of the temperature and polarization pixel windows.

Examples

julia> pixwin(4)
17-element Vector{Float64}:
 1.0000000000020606
 0.9942340766588788
 ⋮
 0.4222841034207188

Healpix.gaussbeam — Function

gaussbeam(fwhm::T, lmax::Int; pol=false) where T

Compute the Gaussian beam window function $B_ℓ$ given the FWHM of the beam in radians, where $C_{ℓ, \mathrm{measured}} = B_ℓ^2 C_ℓ$. This beam is valid in the limit of $σ^2 \ll 0$, which is the case for all high-resolution CMB experiments.

Arguments

fwhm::T: FWHM of the Gaussian beam in radians

lmax::Int: maximum multipole ℓ

pol=false: if false, returns the spin-0 beam for i.e. intensity. if true, returns the spin-2 beam

Returns

Vector{T} containing $B_ℓ$, with the first element referring to ℓ=0.

Algebraic operations in harmonic space

Healpix.jl provides overloads of the Base functions +, -, *, /, as well as LinearAlgebra.dot, allowing to carry out these fundamental operations element-wise in harmonic space directly.

For example, an element-wise sum between two Alm objects can be performed as follows:

using Healpix

#just two constant Alm objects
myalm1 = Healpix.Alm(5,5, ones(ComplexF64, Healpix.numberOfAlms(5)))
myalm1 = Healpix.Alm(5,5, ones(ComplexF64, Healpix.numberOfAlms(5)))

alm_sum = myalm1 + myalm2 #each element will be = 2 + 0im

Multiplying or dividing a set of Alm by a generic function of ℓ or a constant

The operators * and / can be used to multiply or divide an Alm by an $ℓ$-dependent generic function $f_\ell$ (or just a constant, of type Number).

In this case a new instance of Alm type will be returned. To perform a more efficient in-place operation refer to almxfl!, as shown in this brief example:

using Healpix

#just two constant Alm objects
myalm = Healpix.Alm(5,5, ones(ComplexF64, Healpix.numberOfAlms(5)))
myf_l = ones(Healpix.numberOfAlms(5)) .* 2

#will return a new object:
myalm*myf_l
myalm/myf_l

#will overwrite myalm:
almxfl!(myalm, myf_l)
almxfl!(myalm, 1.0 ./ myf_l) #division

In either case the call to such operator consists in a shortcut to the function almxfl.

Healpix.almxfl — Function

almxfl(alm::Alm{Complex{T}}, fl::AbstractVector{T}) where {T <: Number} -> Alm{T}

Multiply an aℓm by a vector bℓ representing an ℓ-dependent function, without changing the a_ℓm passed in input.

Arguments

alms::Alm{Complex{T}}: The Alm object containing the spherical harmonics coefficients

fl::AbstractVector{T}: The array containing the factors fℓ to be multiplied by aℓm

Returns

Alm{Complex{T}}: The result of aℓm * fℓ.

Healpix.almxfl! — Function

almxfl!(alm::Alm{Complex{T}}, fl::AA) where {T <: Number,AA <: AbstractVector{T}}

Multiply IN-PLACE an aℓm by a vector bℓ representing an ℓ-dependent function.

Arguments

alms::Alm{Complex{T}}: The Alm object containing the spherical harmonics coefficients

fl::AbstractVector{T}: The array containing the factors fℓ to be multiplied by aℓm

Base.:+ — Function

+(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number}

Perform the element-wise sum in a_ℓm space. A new Alm object is returned.

Base.:- — Function

-(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number}

Perform the element-wise subtraction in a_ℓm space. A new Alm object is returned.

Base.:* — Function

*(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number}

Perform the element-wise product in a_ℓm space. A new Alm object is returned.

*(alm₁::Alm{Complex{T}}, fl::AA) where {T <: Number, AA <: AbstractVector{T}}

Perform the product of an Alm object by a function of ℓ in a_ℓm space. Note: this consists in a shortcut of almxfl, therefore a new Alm object is returned.

*(alm₁::Alm{Complex{T}}, fl::AbstractVector{T}) where {T <: Number}

Perform the element-wise product of an Alm object by a constant in a_ℓm space.

Base.:/ — Function

/(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number}

Perform an element-wise division in a_ℓm space between two Alms. A new Alm object is returned.

/(alm₁::Alm{Complex{T}}, fl::AA) where {T <: Number,AA <: AbstractVector{T}}

Perform an element-wise division by a function of ℓ in a_ℓm space. A new Alm object is returned.

/(alm₁::Alm{Complex{T}}, c::Number) where {T <: Number}

Perform an element-wise division by a constant in a_ℓm space. A new Alm object is returned.

Dot product

Healpix.jl implements an overload of the operator LinearAlgebra.dot (along with its shortcut ⋅) to perform a dot product directly in harmonic space.

using Healpix
myalm = Healpix.Alm(5,5, ones(ComplexF64, Healpix.numberOfAlms(5)))

dot_res = myalm ⋅ myalm #equivalent to dot(myalm, myalm)

LinearAlgebra.dot — Function

LinearAlgebra.dot(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number}

Implements the dot product in a_ℓm space. The two imput alms must have matching size, lmax and mmax. A new Alm object is returned.

Loading and saving harmonic coefficients

Healpix.readAlmFromFITS — Function

readAlmFromFITS{T <: Complex}(f::CFITSIO.FITSFile, t::Type{T}) -> Alm{T}
readAlmFromFITS{T <: Complex}(fileName::String, t::Type{T}) -> Alm{T}

Read a set of a_ℓm coefficients from a FITS file. If the code fails, CFITSIO will raise an exception. (Refer to the CFITSIO library for more information.)

Healpix.writeAlmToFITS — Function

writeAlmToFITS(f::CFITSIO.FITSFile, alm::Alm{Complex{T}}) where {T <: Number}
writeAlmToFITS(fileName, alm::Alm{Complex{T}}; overwrite = true) where {T <: Number}

Write a set of a_ℓm coefficients into a FITS file. If the code fails, CFITSIO will raise an exception. (Refer to the CFITSIO library for more information.) In the fits file the alms are written with explicit index scheme, $\mathrm{index} = ℓ^2 + ℓ + m + 1$, possibly out of order (check almExplicitIndex).

Alm Indexing

You can use almExplicitIndex to compute the so-called explicit indexing. It is exploited for instance in readAlmFromFITS and writeAlmToFITS.

Healpix.almExplicitIndex — Function

almExplicitIndex(lmax) -> Vector{Int}
almExplicitIndex(lmax, mmax) -> Vector{Int}
almExplicitIndex(alm::Alm{T}) where {T} -> Vector{Int}

Compute the explicit index scheme, i.e. $\mathrm{index} = ℓ^2 + ℓ + m + 1$ up to a certain $ℓ$ and $m$ if specified, or taken from the Alm passed. If not passed, mmax is defaulted to lmax. If lmax and mmax are inconsistent or negative, a DomainError exception is thrown.

Base.eachindex — Function

eachindex(alm::Alm{Complex{T}}) where {T <: Number}

Works as eachindex(alm.alm).

The following functions can be used, in an analogous way as eachindex, in the case of arrays, to obtain sets of indexes or values of $\ell$ and $m$.

On the same line as eachindex, these can be very useful when implementing for-cycles over Alm objects. Here is an example of how to exploit each_ell_m to print explicitly the major-m ordering of a set of complex-stored Alm:

using Random

# Initialize a random set of alm
alm = Alm(4,4, randn(ComplexF64, numberOfAlms(4,4)))

#print a_lm values knowing each corresponding l and m values
function print_alm(alm)
    i=1
    for (l,m) in each_ell_m(alm)
        a_lm = alm.alm[i]
        print("ℓ = $l, |m| = $m: a_ℓm = $a_lm \n")
        i+=1
    end
end

print_alm(alm)

ℓ = 0, |m| = 0: a_ℓm = 0.674166420426957 + 0.12015940852663665im
ℓ = 1, |m| = 0: a_ℓm = 0.26995659480246836 - 0.8844655146931267im
ℓ = 2, |m| = 0: a_ℓm = -0.04981357923916372 + 1.4300721142872432im
ℓ = 3, |m| = 0: a_ℓm = -0.18587542663747117 - 0.8335273998823308im
ℓ = 4, |m| = 0: a_ℓm = -0.09984967044067523 + 0.07514927027045089im
ℓ = 1, |m| = 1: a_ℓm = 1.152496792202116 - 0.17820745395383283im
ℓ = 2, |m| = 1: a_ℓm = -0.6196955065952905 + 0.110454124464739im
ℓ = 3, |m| = 1: a_ℓm = -0.41081488076034667 - 0.17405438966620063im
ℓ = 4, |m| = 1: a_ℓm = 0.1274235417609124 + 0.0007035695547599746im
ℓ = 2, |m| = 2: a_ℓm = 0.4574996495040741 + 0.7724404207390505im
ℓ = 3, |m| = 2: a_ℓm = -0.6792132205217944 + 1.0835084382488063im
ℓ = 4, |m| = 2: a_ℓm = -0.7480159232895027 + 0.38608227767646486im
ℓ = 3, |m| = 3: a_ℓm = -0.8786830655943982 - 1.3432904461294208im
ℓ = 4, |m| = 3: a_ℓm = -0.13105413821878326 - 0.0836995560865279im
ℓ = 4, |m| = 4: a_ℓm = -0.15012829847205797 - 0.6991209480376886im

Healpix.each_ell — Function

each_ell(alm::Alm{Complex{T}}, m::Integer) where {T <: Number} -> Vector{Int}
each_ell(alm::Alm{Complex{T}}, ms::AbstractVector{I}) where {T <: Number, I <: Integer} -> Vector{Int}

Returns an array of all the allowed ℓ values in alm for the given m.

Healpix.each_ell_idx — Function

each_ell_idx(alm::Alm{Complex{T}}, m::Integer) where {T <: Number} -> Vector{Int}
each_ell_idx(alm::Alm{Complex{T}}, ms::AbstractVector{I}) where {T <: Number, I <: Integer} -> Vector{Int}

Returns an array of the indexes of the harmonic coefficients in alm corresponding to all the ℓ values for the given m value(s).

Healpix.each_m — Function

each_m(alm::Alm{Complex{T}}, l::Integer) where {T <: Number} -> Vector{Int}
each_m(alm::Alm{Complex{T}}, ls::AbstractVector{I}) where {T <: Number, I <: Integer} -> Vector{Int}

Returns an array containing all the allowed m values in alm for the given ℓ value(s).

Healpix.each_m_idx — Function

each_m_idx(alm::Alm{Complex{T}}, l::Integer) where {T <: Number} -> Vector{Int}
each_m_idx(alm::Alm{Complex{T}}, ls::AbstractVector{I}) where {T <: Number, I <: Integer} -> Vector{Int}

Returns an array of the indexes of the harmonic coefficients in alm corresponding to all the allowed m values for the given ℓ value(s).

Healpix.each_ell_m — Function

each_ell_m(alm::Alm{Complex{T}}) where {T <: Number} -> Vector{Int}

Returns an array of tuples (l, m) of all the ℓ and m values of alm in m-major order (the same order as how the harmonic coefficients are stored in Alm objects).

Full Pixel Weights

The default map2alm uses iteration to obtain an accurate transform. One can instead apply a pixel weight to compute an accurate transform in a single pass, like quadrature. The easiest way to the pixel weight files is to run

git clone --depth 1 https://github.com/healpy/healpy-data

These weights are in a compressed format that is read with readFullWeights and multiplied into a map with applyFullWeights!.

nside = 32
compressed_weights = Healpix.readFullWeights(
    "healpix_full_weights_nside_$(lpad(nside,4,'0')).fits")
m = Healpix.HealpixMap{Float64,Healpix.RingOrder}(ones(Healpix.nside2npix(nside)))
Healpix.applyFullWeights!(m, compressed_weights)
alm = Healpix.map2alm(m; niter=0)

The subsequent map2alm only needs niter=0.

Healpix.readFullWeights — Function

readFullWeights(filename::String)

Read the set of pixel weights used to compute the generalized Fourier transform of a map.

These weights are usually precomputed; you can download the ones available in the Healpy repository using the following command:

git clone --depth 1 https://github.com/healpy/healpy-data

Arguments:

filename::String: filename of the full pixel weights

Returns:

Vector{Float64}: contains the compressed pixel weights

Healpix.applyFullWeights! — Function

applyFullWeights!(m::HealpixMap{T, RingOrder}, [wgt::Vector{T}]) where T

Apply a pixel weighting to a map for more accurate SHTs. Note that this only helps for lmax<=1.5*Nside. If this is not the case, the pixel weights may do more harm than good.

Pixel weights are automatically downloaded if not specified.

Arguments:

m::HealpixMap{T, RingOrder}: map to modify

wgt::Vector{T} (optional): compressed pixel weights. If not specified, this routine will look for weights in artifacts.

applyFullWeights!(m::PolarizedHealpixMap{T, RingOrder}, [wgt::Vector{T}]) where T

Apply a pixel weighting to a polarized map for more accurate SHTs.

Arguments:

m::PolarizedHealpixMap{T, RingOrder}: map to modify

wgt::Vector{T} (optional): compressed pixel weights. If not specified, an artifact will be sought.