Private API

Healpix.MAX_NSIDE — Constant

const MAX_NSIDE = 2^29

Maximum valid $N_\mathrm{side}$ parameter value.

Healpix.unsafe_ang2pix — Method

p = unsafe_ang2pix(nside, θ, ϕ)

Like ang2pix but neither calls checkhealpix to check the validity of nside nor checks the domain of the spherical coordinates $θ$ and $ϕ$.

Healpix.unsafe_pix2ang — Method

(θ,ϕ) = unsafe_pix2ang(nside, p)

Like pix2ang but does not call checkhealpix to check nside and pixel index validity.

Healpix.unsafe_pix2phi — Method

ϕ = unsafe_pix2phi(nside, p)

Like pix2phi but does not call checkhealpix to check nside and pixel index validity.

Healpix.unsafe_pix2theta — Method

θ = unsafe_pix2theta(nside, p)

Like pix2theta but does not call checkhealpix to check nside and pixel index validity.

Healpix.unsafe_pix2vec — Method

r = unsafe_pix2vec(nside, p)

Like pix2vec but does not call checkhealpix to check nside and pixel index validity.

Healpix.unsafe_pix2z — Method

z = unsafe_pix2z(nside, p)

Like pix2z but does not call checkhealpix to check nside and pixel index validity.

Healpix.unsafe_vec2pix — Method

p = unsafe_vec2pix(nside, r)

Like vec2pix but does not check the validity of the nside or length of $r$.

Healpix.unsafe_zphi2pix — Method

p = unsafe_zphi2pix(nside, z, ϕ)

Like unsafe_ang2pix but takes the value $z = \cos(θ)$ instead of the colatitude $θ$.