Public API

Healpix.Healpix — Module

A module of functions implementing function which interact with the HEALPix pixel definitions. In most cases, only the RING ordering functions are being provided.

See "HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere" Górski, Hivon, & Banday et al (2005) ApJ 622:759–771 arXiv: astro-ph/0409513

Healpix.UNSEEN — Constant

const UNSEEN = -1.6375e+30

Special value recognized by the libhealpix/healpy routines as an unobserved/masked pixel.

Healpix.InvalidNside — Type

InvalidNside(nside)

An invalid nside value was provided.

Healpix.InvalidPixel — Type

InvalidPixel(nside, pix)

An invalid pixel index pix was provided for the given nside.

Healpix.ang2pix — Method

p = ang2pix(nside, θ, ϕ)

Computes the HEALPix pixel index p which contains the point $(θ,ϕ)$ given by the colatitude $θ$ and azimuth $ϕ$, where nside is the Nside resolution factor.

Healpix.checkhealpix — Method

checkhealpix(nside, pix)

Throws an InvalidNside exception if nside is not a valid value or an InvalidPixel exception if pix is out of range for the given $N_\mathrm{side}$.

Healpix.checkhealpix — Method

checkhealpix(nside)

Throws an InvalidNside exception if nside is not a valid value.

Healpix.iscap — Function

iscap(nside, pix)

Test for whether the given pixel pix is in either polar cap for an nside HEALPix map.

Healpix.isequbelt — Function

isequbelt(nside, pix)

Test for whether the given pixel pix is in the equatorial belt for an nside HEALPix map.

Healpix.ishealpixok — Method

isheapixok(nside, pix)

Returns true if nside is valid and pix is in the range 0 to nside2npix(nside) - 1, otherwise false.

Healpix.ishealpixok — Method

ishealpixok(nside)

Returns true if nside is a power of two in the range 1 to 2^29, otherwise false.

Healpix.isnorth — Function

isnorth(nside, pix)

Test for whether the given pixel pix is in the northern hemisphere (including the equatorial ring) for an nside HEALPix map.

Healpix.isnorthcap — Function

isnorthcap(nside, pix)

Test for whether the given pixel pix is in the northern polar cap for an nside HEALPix map.

Healpix.isnorthequbelt — Function

isnorthequbelt(nside, pix)

Test for whether the given pixel pix is in the northern equatorial belt (including the equatorial ring) for an nside HEALPix map.

Healpix.issouth — Function

issouth(nside, pix)

Test for whether the given pixel pix is in the southern hemisphere (excluding the equatorial ring) for an nside HEALPix map.

Healpix.issouthcap — Function

issouthcap(nside, pix)

Test for whether the given pixel pix is in the southern polar cap for an nside HEALPix map.

Healpix.issouthequbelt — Function

issouthequbelt(nside, pix)

Test for whether the given pixel pix is in the southern equatorial belt (excluding the equatorial ring) for an nside HEALPix map.

Healpix.npix2nside — Method

Nisde = npix2nside(npix)

Returns the equivalent Nside corresponding to the number of pixels npix.

Healpix.nring2nside — Method

Nside = nring2nside(nring)

Returns the equivalent Nside corresponding to the number of iso-latitude rings nring.

Healpix.nside2npix — Function

Npix = nside2npix(nside)

Returns the total number of pixels Npix in an nside HEALPix map. Note that HEALPix pixel indexing is 0-based, so valid pixel values are in the range 0 to Npix - 1.

Healpix.nside2npixcap — Function

Npix = nside2npixcap(nside)

Returns the number of pixels Npix in the polar caps for the given nside HEALPix map.

Healpix.nside2npixequ — Function

Npix = nside2npixequ(nside)

Returns the number of pixels Npix in the northern hemisphere, including the equatorial ring, for the given nside HEALPix map.

Healpix.nside2nring — Function

Nring = nside2nring(nside)

Returns the number of iso-latitude rings Nring in the nside HEALPix map.

Healpix.nside2pixarea — Function

σ = nside2pixarea(nside)

Returns the surface area σ (in steradians) of each pixel in the given nside HEALPix map.

Healpix.pix2ang — Method

(θ,ϕ) = pix2ang(nside, p)

Computes the colatitude and azimuth pair $(θ,ϕ)$ for the given pixel p, where nside is the Nside resolution factor.

Healpix.pix2phi — Method

ϕ = pix2phi(nside, p)

Computes the azimuth $ϕ$ for the given pixel p, where nside is the Nside resolution factor.

Healpix.pix2ring — Method

i = pix2ring(nside, p)

Computes the ring index i for the given pixel p, where nside is the Nside resolution factor.

Healpix.pix2ringidx — Method

j = pix2ringidx(nside, p)

Computes the index j within the ring for the given pixel p, where nside is the Nside resolution factor.

Healpix.pix2theta — Method

θ = pix2theta(nside, p)

Computes the colatitude $θ$ for the given pixel p, where nside is the Nside resolution factor.

Healpix.pix2vec — Method

r = pix2vec(nside, p)

Computes the unit vector $r$ pointing to the pixel center of the given pixel p, where nside is the Nside resolution factor.

Healpix.pix2z — Method

z = pix2z(nside, p)

Computes the cosine of the colatitude $z = \cos(θ)$ for the given pixel p, where nside is the Nside resolution factor.

Healpix.vec2pix — Method

p = vec2pix(nside, r)

Computes the HEALPix pixel index p which contains the point at the end of the unit vector $r$, where nside is the Nside resolution factor.