Miscellanea

General-purpose functions

Healpix.jl implements a few generic functions that can be helpful when doing calculations on the sphere.

Healpix.normalizeAngle — Method.

normalizeAngle(x)

Return the same angle as the argument, but in the range [0, 2π). Note that this is slightly different from mod2pi, as the latter returns a value in the range [0, 2π].

source

Healpix.lat2colat — Method.

lat2colat(x)

Convert latitude into colatitude. Both x and the result are expressed in radians.

source