Power Spectrum

Power spectrum components $C_{\ell}$ are encoded as Vector{T}. You can use the function alm2cl to convert a set of $a_{\ell m}$ coefficients into the components $C_\ell$ of the power spectrum.

Loading and saving power spectrum components

Healpix.jl implements functions to read/write the components $C_{\ell}$ from/to a FITS files.

Healpix.readClFromFITS — Function

readClFromFITS{T <: Real}(f::CFITSIO.FITSFile, t::Type{T}; col_num = 2) -> Vector{T}
readClFromFITS{T <: Real}(fileName::String, t::Type{T}; col_num = 2) -> Vector{T}

Read a set of C_ℓ coefficients from a FITS file.

Missing docstring.

Missing docstring for writeClToFITS. Check Documenter's build log for details.

Converting different power spectrum representation

It's often useful to represent, especially for plotting it, the power spectrum in the form of $D_{\ell}$. Healpix.jl implements a couple of functions to convert a power spectrum from/to such a representation.

Missing docstring.

Missing docstring for cl2dl. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dl2cl. Check Documenter's build log for details.