Spherical harmonics

Starting from version 2.4, Healpix.jl implements generalized Fourier transformations through the libsharp library, to convert a map from its pixel-space representation to its decomposition in spherical harmonics. This has multiple applications, the most relevant being the analysis of Cosmic Microwave Background maps and the efficient computation of convolution operators.

Everything revolves around the Alm type, which encodes a set of spherical harmonics and is thus conceptually equivalent to the concept of a Map, only living in the harmonic space:

Healpix.Alm — Type

An array of harmonic coefficients (a_ℓm).

The type T is used for the value of each harmonic coefficient, and it must be a Number (one should however only use complex types for this). The type AA is used to store the array of coefficients; a typical choice is Vector.

A Alm type contains the following fields:

alm: the array of harmonic coefficients

lmax: the maximum value for $ℓ$

mmax: the maximum value for $m$

tval: maximum number of $m$ coefficients for the maximum $ℓ$

In the general case, the number of coefficients in a spherical harmonic expansion is infinite. For obvious reasons, Healpix.jl only allows to store band-limited expansions. The function numberOfAlms returns the number of floating-point numbers used to store the expansion, as a function of the maximum value for $\ell$ and $m$.

Healpix.numberOfAlms — Function

numberOfAlms(lmax::Integer, mmax::Integer) -> Integer
numberOfAlms(lmax::Integer) -> Integer

Return the size of the array of complex numbers needed to store the a_lm coefficients in the range of ℓ and m specified by lmax and mmax. If mmax is not specified, it is assumed to be equal to lmax. If lmax and mmax are inconsistent or negative, a DomainError exception is thrown.

Converting between pixel space and harmonic space

Healpix.jl implements the four functions alm2map, map2alm, alm2map!, and map2alm! to convert a map from a pixel-space representation to the harmonic space and vice-versa. The functions ending with ! are mutating functions, which means that they assume that the result must be saved in a preallocated variable; they are space- and time-efficient and should be used when you want your code to be performant, or when you plan to apply the same operation several times (e.g., in a Monte Carlo simulation).

Here is an example:

using Random

# Ensure reproducibility by using a fixed seed
Random.seed!(1234)

nside = 8
m = Map{Float32,RingOrder}(nside)

# Initialize the pixels to random values in the 0…1 range
for i in 1:length(m)
    m[i] = rand(Float32)
end

alm = map2alm(m)

# Go back to pixel space
newm = alm2map(alm, nside)

768-element Map{Float64, RingOrder, Vector{Float64}}:
  0.1652631874267924
  0.3618781044964514
 -0.0013815388023583064
  0.40084766002795
  0.2530007233310308
  0.7785956979193502
  0.43071187199759775
  0.9379452241907033
  0.7554174116102925
  0.6032735376338993
  ⋮
  0.44347259128908373
  0.42263793922526627
 -0.20520639215243272
  0.4165901675651938
  0.6860000214208772
  0.7211351759306383
  0.6707478409770005
  0.47791307158759333
  0.7109101053740821

The variable newm is a map that is close enough to m, yet it is not exactly the same because of the approximations done by both map2alm and alm2map.

Healpix.map2alm! — Function

map2alm!(map::Map{Float64, RingOrder, Array{Float64, 1}}, alm::Alm{ComplexF64, Array{ComplexF64, 1}}; niter::Integer=3)
map2alm!(map::PolarizedMap{Float64, RingOrder, Array{Float64, 1}}, alm::Array{Alm{ComplexF64, Array{ComplexF64, 1}},1};
    niter::Integer=3)

This function performs a spherical harmonic transform on the map and places the results in the passed alm object. This function requires types derived from Float64, since it is done in-place.

Arguments

map: the map that must be decomposed in spherical harmonics. It can either be a Map{Float64, RingOrder} type (scalar map) or a PolarizedMap{Float64, RingOrder} type (polarized map).

alm::Alm{ComplexF64, Array{ComplexF64, 1}}: the spherical harmonic coefficients to be written to.

Keywords

niter::Integer: number of iterations of SHTs to perform, to enhance accuracy

Healpix.map2alm — Function

map2alm(map::Map{Float64, RingOrder, AA};
    lmax=nothing, mmax=nothing, niter::Integer=3)
map2alm(m::Map{T, RingOrder, AA}; lmax=nothing, mmax=nothing,
    niter::Integer=3) where {T <: Real, AA <: AbstractArray{T, 1} }

Compute the spherical harmonic coefficients of a map. To enhance precision, more iterations of the transforms can be performed by passing a nonzero niter. The underlying SHT library libsharp performs all calculations using Cdouble types, so all inputs are converted to types based on Float64.

Arguments

map: the map to decompose in spherical harmonics. It can either be a Map{T, RingOrder, AA} type (scalar map) or a PolarizedMap{T, RingOrder, AA} type (polarized map).

Keywords

lmax::Integer: the maximum ℓ coefficient, will default to 3*nside-1 if not specified.

mmax::Integer: the maximum m coefficient

niter::Integer: number of SHT iterations, to enhance precision. Defaults to 3

Returns

Alm{ComplexF64, Array{ComplexF64, 1}}: the spherical harmonic coefficients corresponding to the map

Healpix.alm2map! — Function

alm2map!(alm::Alm{ComplexF64, Array{ComplexF64, 1}}, map::Map{Float64, RingOrder, Array{Float64, 1}})
alm2map!(alm::Array{Alm{ComplexF64, Array{ComplexF64, 1}},1}, map::PolarizedMap{Float64, RingOrder, Array{Float64, 1}})

This function performs a spherical harmonic transform on the map and places the results in the passed alm object. This function requires types derived from Float64, since it is done in-place.

Arguments

alm::Alm{ComplexF64, Array{ComplexF64, 1}}: the spherical harmonic coefficients to perform the spherical harmonic transform on.

map: the map that will contain the result. It can either be a Map{Float64, RingOrder, Array{Float64, 1}} type (scalar map) or a PolarizedMap{Float64, RingOrder, Array{Float64, 1}} (polarized map).

Healpix.alm2map — Function

alm2map(alm::Alm{ComplexF64, Array{Float64, 1}}, nside::Integer)
alm2map(alm::Alm{T, Array{T, 1}}, nside::Integer) where T
alm2map(alm::Array{Alm{ComplexF64, Array{ComplexF64, 1}},1}, nside::Integer)
alm2map(alms::Array{Alm{T, Array{T, 1}},1}, nside::Integer) where T

Compute a map from spherical harmonic coefficients. The underlying SHT library libsharp performs all calculations in Cdouble, so all inputs are converted to types based on Float64.

Arguments

alm: the spherical harmonic coefficients to transform. If of type Alm{T, Array{T, 1}}, we assume a spin-0 spherical harmonic transform. If an array of Alm is passed, we assume that the components correspond to T, E, and B coefficients.

Keywords

nside::Integer: Healpix resolution parameter

Returns

Map{Float64, RingOrder, Array{Float64, 1}} or PolarizedMap{Float64, RingOrder, Array{Float64, 1}} depending on if the input alm is of type Alm{T, Array{T, 1}} or Array{Alm{T, Array{T, 1}}} respectively.

From harmonic coefficients to the power spectrum

You can use the function alm2cl to convert a set of $a_{\ell m}$ coefficients into the components $C_\ell$ of the power spectrum.

Healpix.alm2cl — Function

alm2cl(alm::Alm{Complex{T}}) where {T <: Number}
alm2cl(alm₁::Alm{Complex{T}}, alm₂::Alm{Complex{T}}) where {T <: Number}

Compute $C_{\ell}$ from the spherical harmonic coefficients of one or two fields.

Arguments

alm₁::Alm{Complex{T}}: the spherical harmonic coefficients of the first field

alm₂::Alm{Complex{T}}: the spherical harmonic coefficients of the second field

Returns

Array{T} containing C_{ll}, with the first element referring to ℓ=0.

Loading and saving harmonic coefficients

Healpix.readAlmFromFITS — Function

readAlmFromFITS{T <: Complex}(f::FITSIO.FITSFile, t::Type{T}) -> Alm{T}
readAlmFromFITS{T <: Complex}(fileName::String, t::Type{T}) -> Alm{T}

Read a set of a_ℓm coefficients from a FITS file. If the code fails, FITSIO will raise an exception. (Refer to the FITSIO library for more information.)

Full Pixel Weights

The default map2alm uses iteration to obtain an accurate transform. One can instead apply a pixel weight to compute an accurate transform in a single pass, like quadrature. The easiest way to the pixel weight files is to run

git clone --depth 1 https://github.com/healpy/healpy-data

These weights are in a compressed format that is read with readfullweights and multiplied into a map with applyweights!.

nside = 32
compressed_weights = Healpix.readfullweights(
    "healpix_full_weights_nside_$(lpad(nside,4,'0')).fits")
m = Healpix.Map{Float64,Healpix.RingOrder}(ones(Healpix.nside2npix(nside)))
Healpix.applyweights!(m, compressed_weights)
alm = Healpix.map2alm(m; niter=0)

The subsequent map2alm only needs niter=0.

Healpix.readfullweights — Function

readfullweights(filename::String)

The easiest way to the pixel weight files is to run

git clone --depth 1 https://github.com/healpy/healpy-data

Arguments:

filename::String: filename of the full pixel weights

Returns:

Vector{Float64}: contains the compressed pixel weights

Healpix.applyweights! — Function

applyweights!(m::Map{T, RingOrder}, wgt::Vector{T}) where T

Apply a pixel weighting to a map for more accurate SHTs.

Arguments:

m::Map{T, RingOrder}: map to modify

wgt::Vector{T}: compressed pixel weights

applyweights!(m::PolarizedMap{T, RingOrder}, wgt::Vector{T}) where T

Apply a pixel weighting to a polarized map for more accurate SHTs.

Arguments:

m::PolarizedMap{T, RingOrder}: map to modify

wgt::Vector{T}: compressed pixel weights