Spherical harmonics

Spherical harmonics

The support for spherical harmonics in Healpix.jl is still woefully inadequate. Only a few functions to load and store harmonic coefficients are available. Everything revolves around the Alm type:

Healpix.Alm — Type.

An array of a_ℓm numbers.

source

The number of coefficients in a spherical harmonic expansion is infinite. For obvious reasons, Healpix.jl only allows to store band-limited expansions. The function numberOfAlms returns the number of floating-point numbers used to store the expansion, as a function of the maximum value for $\ell$ and $m$.

Healpix.numberOfAlms — Function.

numberOfAlms(lmax::Integer, mmax::Integer) -> Integer
numberOfAlms(lmax::Integer) -> Integer

Return the size of the array of complex numbers needed to store the a_lm coefficients in the range of ℓ and m specified by lmax and mmax. If mmax is not specified, it is assumed to be equal to lmax. If lmax and mmax are inconsistent or negative, a DomainError exception is thrown.

source

Loading and saving harmonic coefficients

Healpix.readAlmFromFITS — Function.

readAlmFromFITS{T <: Complex}(f::FITSIO.FITSFile, t::Type{T}) -> Alm{T}
readAlmFromFITS{T <: Complex}(fileName::String, t::Type{T}) -> Alm{T}

Read a set of a_ℓm coefficients from a FITS file. If the code fails, FITSIO will raise an exception. (Refer to the FITSIO library for more information.)

source