Miscellanea

General-purpose functions

Healpix.jl implements a few generic functions that can be helpful when doing calculations on the sphere.

Healpix.lat2colat — Function.

lat2colat(x) colat2lat(x)

Convert colatitude into latitude and vice versa. Both x and the result are expressed in radians.

source

Healpix.colat2lat — Function.

lat2colat(x) colat2lat(x)

Convert colatitude into latitude and vice versa. Both x and the result are expressed in radians.

source