Visualization

Visualization functions

Healpix.jl implements the project function, which creates a 2D matrix containing the cartographic projection of a map. A few standard cartographic projections are implemented, but users can provide their own projections. The function can optionally use heatmap (from the Plots.jl package) to display the 2D matrix. Two useful wrappers to project are equirectangular and mollweide, which employ the equirectangular and Mollweide projections respectively.

using Healpix

nside = 8
m = Map{Float64, RingOrder}(nside)
m.pixels[:] = 1:length(m.pixels)
mollweide(m)

Healpix.project — Function.

project(m::Map{T, O}; kwargs...) where {T <: Number, O <: Order}

Return a 2D bitmap (array) containing a cartographic projection of the map and a 2D bitmap containing a boolean mask. The size of the bitmap is specified by figsize, which must be a 2-tuple. The function projfn must be a function which accepts as input two parameters x and y (numbers between -1 and 1).

The following keywords can be used in the call:

figsize: 2-tuple specifying the (height, width) of the bitmap in pixels

center: 2-tuple specifying the location (colatitude, longitude) of the sky point that is to be placed in the middle of the image (in radians)

show: Boolean; if true (the default), the bitmap will be shown using functions from the "Plots" package.

returnmask: Boolean; if true, the function returns a 2-tuple containing the image bitmap and a mask bitmap, which is set to true if the pixel falls within the carthographic projection, false otherwise. If returnmask is false, only the image bitmap is returned.

show: Boolean. If true (the default), the map will be displayed. It has no effect if returnmask is true.

source

Cartographic projections

Healpix.equiprojinv — Function.

function equiprojinv(x, y)

Inverse equirectangular projection. Given a point (x, y) on the plane [-1, 1] × [-1, 1], return a tuple (Bool, Number, Number) where the first Boolean is a flag telling if the point falls within the projection (true) or not (false), and the two numbers are the latitude and colatitude in radians.

source

Healpix.mollweideprojinv — Function.

function mollweideprojinv(x, y)

Inverse Mollweide projection. Given a point (x, y) on the plane, with x ∈ [-1, 1], y ∈ [-1, 1], return a 3-tuple of type (Bool, Number, Number). The boolean specifies if (x, y) falls within the map (true) or not (false), the second and third arguments are the latitude and longitude in radians.

source

High-level wrappers to project

Healpix.equirectangular — Function.

equirectangular(m::Map{T,O}; kwargs...) where {T <: Number, O <: Order}

High-level wrapper around project for equirectangular projections.

source

Healpix.mollweide — Function.

mollweide(m::Map{T,O}; kwargs...) where {T <: Number, O <: Order}

High-level wrapper around project for Mollweide projections.

source